git: get info from a 'bare' repo instead
For git repositories you basically always pay the storage cost twice:
once for the tree in .git, once for the actual files being checked out.
Additionally, the tree in .git is compressed, so in many cases it can be
smaller than the actual files being checked out.

This commit updates Hound to optionally process git files from the
actual git files instead of the checked out tree.

For me, this saves a significant amount of disk space. Why does it
matter? Because I run hound on a laptop instead of a datacenter, and
space is limited. It does cost a few more CPU cycles, but that's fine by
me.